Refracting code and modularizing it
